// src/app/shared/enum/operation.enum.ts
export enum OperationEnum
{
    CREATE = "create",
    EDIT = "edit",
    DELETE = "delete",
    EDIT_PARENT = "edit-parent",
    EDIT_CHILD = "edit-child",
    EDIT_SUB_CHILD = "edit-sub-child",
    DELETE_PARENT = "delete-parent",
    DELETE_CHILD = "delete-child",
    DELETE_SUB_CHILD = "delete-sub-child",
}