// src/app/shared/model/sub-child-menu.model.ts
/**
 * © Rolling Array https://rollingarray.co.in/
 *
 *
 * @summary Sub Child menu model
 * @author code@rollingarray.co.in
 *
 * Created at     : 2021-12-16 20:52:39 
 * Last modified  : 2021-12-16 20:52:39 
 */


export interface SubChildMenuModel
{
    articleId: string;
    articleTitle: string;
    subChildMenuOrder: number;
}